qa/rag/requirements.txt
markdownify
trafilatura
pandas
beautifulsoup4
agentops
fastapi
uvicorn
requests
pydantic-settings